ENH: update tank inertia evaluation for non flat caps and gaseous portion.
