+++ rocketpy/motors/__init__.py
@@ -1,6 +1,6 @@
 from .Motor import Motor
 from .Fluid import Fluid
+from .Tank import Tank, MassFlowRateBasedTank, UllageBasedTank, MassBasedTank
 from .SolidMotor import SolidMotor
 from .LiquidMotor import LiquidMotor
-from .LiquidMotor import Tank, MassFlowRateBasedTank, UllageBasedTank, MassBasedTank
 from .HybridMotor import HybridMotor
